requirements.txt
streamlit
python-dotenv
PyPDF2
python-docx
langchain
chromadb
sentence-transformers
numpy
ollama
docx
